feat(vim): implemented vim's packages feature
Clone (submodules) to .vim/pack/bundle/opt. When a plugin resides in
this directory in can be loaded in vim with either `packadd` or
`packadd!`. The bang causes vim to only load the matching directory and
NOT load any plugin files or ftdetect scripts. Instead, these will be
loaded during initialization