feat(nvim-lsp): first setup
Added the nvim-lsp package and configured it to be used for LSP
features. An additional package was required to get it to show in vims
omni completion menu; `completion-nvim`. I load this package with
packadd only when it is needed in the lsp_config.lua.

TODO: commented out the line where completion-nvim will use ulti snips
because I haven't implemented that yet

I am unsure whether the completion-nvim package is better then
deoplete-nvim. Might switch to that one later on, not sure yet.