feat(vim): added mappings for git
Some sane mappings for fugitive interaction. Also added a mapping to
clear the search highlighting with escape in normal mode (since esq is
not used for anything else). Finally a mapping for the search mapping
'/' to add a '\v' when searching. This causes the cursor to become
visible again of disappearing when tapping the search key