fix(vim colors): wrapped setting highlights in func
Some highlight groups were not applied whilst others where. Wrapping
them in a function and calling the function from an augroup seems to fix
this issue. Though I am not entirely sure yet, I believe this is caused
by commands outside of functions/augroups only getting evaluated once
(and can therefore be overwritten).

Will have to research this a bit more later on. For now it is fixed