feat(vim): textwidth and colorcolumn
Textwidth is set to 80 which means that insert mode will start breaking
the line and continue on a new line when typing more than 80 characters.
The nice thing is, as far as I know, it only breaks for comments and
markdown and not for code.

ColorColumn is set to +1, which means textwidth +1. This will draw a
vertical bar in the buffer to indicate where the 80 col textwidth is.
Makes it easier to know when to break of code or refactor some code
because it is gettings to long.