feat(vim): statusline
This is very heavily based on Wincent's statusline.. I do understand
what is happening and how it works. For now it is fine by me, but later
on I would like to give my own twist to it and really create my own
statusline instead of copying from someone else.